--- packages/launchpad/src/migration/ChangeExtensionModalStep1.cy.tsx
@@ -1,7 +1,0 @@
-import ChangeExtensionModalStep1 from './ChangeExtensionModalStep1.vue'
-
-describe('<ChangeExtensionModalStep1/>', { viewportWidth: 1119 }, () => {
-  it('renders expected content', () => {
-    cy.mount(ChangeExtensionModalStep1)
-  })
-})
