--- packages/launchpad/src/migration/ChangeExtensionModalStep2.cy.tsx
@@ -1,7 +1,0 @@
-import ChangeExtensionModalStep2 from './ChangeExtensionModalStep2.vue'
-
-describe('<ChangeExtensionModalStep2/>', { viewportWidth: 1119 }, () => {
-  it('renders expected content', () => {
-    cy.mount(ChangeExtensionModalStep2)
-  })
-})
