+++ packages/launchpad/src/migration/OptOutModalStep1.cy.tsx
@@ -1,0 +1,7 @@
+import OptOutModalStep1 from './OptOutModalStep1.vue'
+
+describe('<OptOutModalStep1/>', { viewportWidth: 1119 }, () => {
+  it('renders expected content', () => {
+    cy.mount(OptOutModalStep1)
+  })
+})
