+++ packages/launchpad/src/migration/OptOutModalStep2.cy.tsx
@@ -1,0 +1,7 @@
+import OptOutModalStep2 from './OptOutModalStep2.vue'
+
+describe('<OptOutModalStep2/>', { viewportWidth: 1119 }, () => {
+  it('renders expected content', () => {
+    cy.mount(OptOutModalStep2)
+  })
+})
